# Avoid false sharing on thread pool data structures