--- a/onnxruntime/test/onnx/microbenchmark/tptest.cc
+++ b/onnxruntime/test/onnx/microbenchmark/tptest.cc
@@ -1,4 +1,3 @@
-#include <iostream>
 #include <benchmark/benchmark.h>
 #include <core/platform/threadpool.h>
 #include <core/util/thread_utils.h>